pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "privaterag"
version = "0.1.0"
description = "Topic-driven automatic collection, validation, and RAG search system"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0,<0.112",
    "uvicorn[standard]>=0.29.0,<0.30",
    # Database
    "sqlalchemy>=2.0.30,<3.0",
    "aiosqlite>=0.20.0,<0.21",
    "alembic>=1.13.0,<2.0",
    # Validation & Settings
    "pydantic>=2.7.0,<3.0",
    "pydantic-settings>=2.3.0,<3.0",
    "pyyaml>=6.0.1",
    # Vector DB
    "qdrant-client>=1.9.0,<2.0",
    # Embeddings
    "sentence-transformers>=3.0.0,<4.0",
    # LLM Providers
    "httpx>=0.27.0,<0.28",
    "openai>=1.30.0,<2.0",
    "anthropic>=0.28.0,<0.29",
    "google-generativeai>=0.7.0,<0.8",
    # Scheduler
    "apscheduler>=3.10.4,<4.0",
    # Content extraction
    "trafilatura>=1.9.0,<2.0",
    "readability-lxml>=0.8.1",
    "lxml>=5.2.0,<6.0",
    "playwright>=1.44.0,<2.0",
    # Retry logic
    "tenacity>=8.3.0,<9.0",
    # Language detection
    "langdetect>=1.0.9",
    # Text splitting
    "langchain-text-splitters>=0.2.0,<0.3",
    # Async file I/O
    "aiofiles>=23.2.1,<24.0",
    # Logging
    "structlog>=24.2.0,<25.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0",
    "pytest-asyncio>=0.23.0,<0.24",
    "pytest-cov>=5.0.0,<6.0",
    "pytest-mock>=3.14.0,<4.0",
    "freezegun>=1.5.0,<2.0",
    "httpx>=0.27.0",   # for TestClient
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: 단위 테스트",
    "integration: 통합 테스트",
    "api: API 엔드포인트 테스트",
    "slow: 실행 시간이 긴 테스트 (E2E 등)",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.coverage.run]
source = ["app"]
omit = ["app/db/migrations/*", "tests/*"]
